Add casting to pql() query, refactor more
and fix #202
